--- a/hystrix/pool_test.go
+++ b/hystrix/pool_test.go
@@ -1,10 +1,9 @@
 package hystrix
 
 import (
+	"sync/atomic"
 	"testing"
 	"time"
-
-	"sync/atomic"
 
 	. "github.com/smartystreets/goconvey/convey"
 )