--- a/hystrix/rolling/rolling_timing.go
+++ b/hystrix/rolling/rolling_timing.go
@@ -7,6 +7,15 @@ import (
 	"time"
 )
 
+const (
+	// This is the number of items we search for or sum (e.g. last 60 seconds)
+	timingWindow = int64(60)
+
+	// This is the number of items used to store data.
+	// The extra item is stop collision when the window "wraps around" to the next second.
+	timingItems = timingWindow + 1
+)
+
 // Timing maintains time Durations for each time bucket.
 // The Durations are kept in an array to allow for a variety of
 // statistics to be calculated from the source data.
@@ -16,18 +25,36 @@ type Timing struct {
 
 	CachedSortedDurations []time.Duration
 	LastCachedTime        int64
+
+	// allow of mocking of time in tests
+	timeGeneratorSec  func() int64
+	timeGeneratorNano func() int64
 }
 
 type timingBucket struct {
+	timestamp int64
 	Durations []time.Duration
+}
+
+// reset/empty the bucket
+func (t *timingBucket) empty() {
+	t.timestamp = 0
+	// is there something better than this?
+	t.Durations = nil
 }
 
 // NewTiming creates a RollingTiming struct.
 func NewTiming() *Timing {
 	r := &Timing{
-		Buckets: make(map[int64]*timingBucket),
+		Buckets: make(map[int64]*timingBucket, timingWindow+1),
 		Mutex:   &sync.RWMutex{},
 	}
+
+	// create all the buckets
+	for x := int64(0); x < timingItems; x++ {
+		r.Buckets[x] = &timingBucket{}
+	}
+
 	return r
 }
 
@@ -44,20 +71,22 @@ func (r *Timing) SortedDurations() []time.Duration {
 	t := r.LastCachedTime
 	r.Mutex.RUnlock()
 
-	if t+time.Second.Nanoseconds() > time.Now().UnixNano() {
+	now := time.Now()
+	nowNano := r.getTimeInNano(now)
+	if t+time.Second.Nanoseconds() > nowNano {
 		// don't recalculate if current cache is still fresh
 		return r.CachedSortedDurations
 	}
 
 	var durations byDuration
-	now := time.Now()
+	minTimeInSec := r.getMinTimeInSec(r.getTimeInSec(now))
 
 	r.Mutex.Lock()
 	defer r.Mutex.Unlock()
 
-	for timestamp, b := range r.Buckets {
-		// TODO: configurable rolling window
-		if timestamp >= now.Unix()-60 {
+	var b *timingBucket
+	for _, b = range r.Buckets {
+		if b.timestamp >= minTimeInSec {
 			for _, d := range b.Durations {
 				durations = append(durations, d)
 			}
@@ -67,48 +96,27 @@ func (r *Timing) SortedDurations() []time.Duration {
 	sort.Sort(durations)
 
 	r.CachedSortedDurations = durations
-	r.LastCachedTime = time.Now().UnixNano()
+	r.LastCachedTime = nowNano
 
 	return r.CachedSortedDurations
 }
 
-func (r *Timing) getCurrentBucket() *timingBucket {
-	r.Mutex.RLock()
-	now := time.Now()
-	bucket, exists := r.Buckets[now.Unix()]
-	r.Mutex.RUnlock()
-
-	if !exists {
-		r.Mutex.Lock()
-		defer r.Mutex.Unlock()
-
-		r.Buckets[now.Unix()] = &timingBucket{}
-		bucket = r.Buckets[now.Unix()]
-	}
-
-	return bucket
-}
-
-func (r *Timing) removeOldBuckets() {
-	now := time.Now()
-
-	for timestamp := range r.Buckets {
-		// TODO: configurable rolling window
-		if timestamp <= now.Unix()-60 {
-			delete(r.Buckets, timestamp)
-		}
-	}
-}
-
 // Add appends the time.Duration given to the current time bucket.
 func (r *Timing) Add(duration time.Duration) {
-	b := r.getCurrentBucket()
-
 	r.Mutex.Lock()
 	defer r.Mutex.Unlock()
 
+	timeInSec := r.getTimeInSec(time.Now())
+	index := r.getIndex(timeInSec)
+
+	b := r.Buckets[index]
+	if b.timestamp != timeInSec {
+		// auto-empty buckets that are not clean (caused by sporadic data)
+		b.empty()
+		b.timestamp = timeInSec
+	}
+
 	b.Durations = append(b.Durations, duration)
-	r.removeOldBuckets()
 }
 
 // Percentile computes the percentile given with a linear interpolation.
@@ -144,5 +152,29 @@ func (r *Timing) Mean() uint32 {
 		return 0
 	}
 
-	return uint32(sum.Nanoseconds()/length) / 1000000
+	return uint32(sum.Nanoseconds() / length / 1000000)
+}
+
+func (r *Timing) getTimeInSec(now time.Time) int64 {
+	if r.timeGeneratorSec != nil {
+		return r.timeGeneratorSec()
+	}
+
+	return now.Unix()
+}
+
+func (r *Timing) getTimeInNano(now time.Time) int64 {
+	if r.timeGeneratorNano != nil {
+		return r.timeGeneratorNano()
+	}
+
+	return now.UnixNano()
+}
+
+func (r *Timing) getMinTimeInSec(timeInSec int64) int64 {
+	return timeInSec - timingWindow + 1
+}
+
+func (r *Timing) getIndex(timeInSec int64) int64 {
+	return timeInSec % timingItems
 }