--- a/hystrix/rolling/rolling_timing_test.go
+++ b/hystrix/rolling/rolling_timing_test.go
@@ -1,10 +1,12 @@
 package rolling
 
 import (
+	"fmt"
 	"testing"
 	"time"
 
 	. "github.com/smartystreets/goconvey/convey"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestOrdinal(t *testing.T) {
@@ -67,4 +69,101 @@ func TestOrdinal(t *testing.T) {
 			})
 		})
 	})
+}
+func TestTiming_2seconds(t *testing.T) {
+	timing := NewTiming()
+
+	// time generator (predictable time)
+	// This one creates 2 consecutive buckets
+	calls := int64(0)
+	timing.timeGeneratorSec = func() int64 {
+		calls++
+		if calls <= 100 {
+			return int64(61)
+		}
+		return int64(62)
+	}
+
+	timing.timeGeneratorNano = func() int64 {
+		return timing.timeGeneratorSec() * time.Second.Nanoseconds()
+	}
+
+	// call add
+	for x := 0; x < 150; x++ {
+		timing.Add(1 * time.Second)
+	}
+
+	// validate
+	assert.Equal(t, uint32(1*time.Second/1000000), timing.Percentile(1))
+	assert.Equal(t, uint32(1*time.Second/1000000), timing.Percentile(25))
+	assert.Equal(t, uint32(1*time.Second/1000000), timing.Percentile(50))
+	assert.Equal(t, uint32(1*time.Second/1000000), timing.Percentile(75))
+	assert.Equal(t, uint32(1*time.Second/1000000), timing.Percentile(99))
+	assert.Equal(t, uint32(1*time.Second/1000000), timing.Mean())
+}
+
+func TestTiming_60seconds(t *testing.T) {
+	timing := NewTiming()
+
+	// time generator (predictable time)
+	// This one creates 2 consecutive buckets
+	calls := int64(0)
+	timing.timeGeneratorSec = func() int64 {
+		calls++
+		if calls <= 60 {
+			return 60 + calls
+		}
+		return 60 + 60
+	}
+
+	timing.timeGeneratorNano = func() int64 {
+		return timing.timeGeneratorSec() * time.Second.Nanoseconds()
+	}
+
+	// call add
+	for x := 0; x < 60; x++ {
+		timing.Add(time.Duration(1+x) * time.Second)
+	}
+
+	// validate
+	assert.Equal(t, uint32(1*time.Second/1000000), timing.Percentile(1))
+	assert.Equal(t, uint32(15*time.Second/1000000), timing.Percentile(25))
+	assert.Equal(t, uint32(30*time.Second/1000000), timing.Percentile(50))
+	assert.Equal(t, uint32(45*time.Second/1000000), timing.Percentile(75))
+	assert.Equal(t, uint32(60*time.Second/1000000), timing.Percentile(99))
+	// mean of 1,2,...,60
+	assert.Equal(t, uint32(float64(30.5)*float64(time.Second)/1000000), timing.Mean())
+}
+
+func TestTiming_100seconds(t *testing.T) {
+	timing := NewTiming()
+
+	// time generator (predictable time)
+	// This one creates 2 consecutive buckets
+	calls := int64(0)
+	timing.timeGeneratorSec = func() int64 {
+		calls++
+		if calls <= 100 {
+			return 60 + calls
+		}
+		return 60 + 100
+	}
+
+	timing.timeGeneratorNano = func() int64 {
+		return timing.timeGeneratorSec() * time.Second.Nanoseconds()
+	}
+
+	// call add
+	for x := 0; x < 100; x++ {
+		timing.Add(time.Duration(1+x) * time.Second)
+	}
+
+	// validate
+	assert.Equal(t, uint32(41*time.Second/1000000), timing.Percentile(1), fmt.Sprintf("expected %.3f; was %.3f", float64(41*time.Second/1000000), float64(timing.Percentile(1))))
+	assert.Equal(t, uint32(55*time.Second/1000000), timing.Percentile(25))
+	assert.Equal(t, uint32(70*time.Second/1000000), timing.Percentile(50))
+	assert.Equal(t, uint32(85*time.Second/1000000), timing.Percentile(75))
+	assert.Equal(t, uint32(100*time.Second/1000000), timing.Percentile(99))
+	// mean of 41,42,...,100
+	assert.Equal(t, uint32(float64(70.5)*float64(time.Second)/1000000), timing.Mean(), fmt.Sprintf("expected %.3f; was %.3f", float64(70.5)*float64(time.Second/1000000), float64(timing.Mean())))
 }