--- a/plugins/graphite_aggregator.go
+++ b/plugins/graphite_aggregator.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/myteksi/hystrix-go/hystrix/metric_collector"
+	metrics "github.com/rcrowley/go-metrics"
 )
 
 var makeTimerFunc = func() interface{} { return metrics.NewTimer() }