Change statistic bucketing from timestamp based, to a rolling window.

This reduces the memory churn (allocations and GC) and according to benchmarks improves speed as well

Bench Compare

	benchmark                             old ns/op     new ns/op     delta
	BenchmarkRollingNumberIncrement-8     220           85.5          -61.14%

The original discussion was here: https://diff.myteksi.net/D84737
Includes #16 and #15